🗃Refactor: OptionGroups 분리
