+++ src/domain/index.ts
@@ -6,7 +6,9 @@ export * from '@/domain/beverage/Material';
 export * from '@/domain/beverage/Recipe';
 export * from '@/domain/beverage/Beverage';
 
-export * from '@/domain/option/OptionGroup';
+export * from '@/domain/option/group/OptionGroup';
+export * from '@/domain/option/group/OptionGroups';
+
 export * from '@/domain/option/Option';
 export * from '@/domain/option/Options';
